made all changes according to review
